Fix int to uint64

+++ dice_roller/roller.go
@@ -5,7 +5,7 @@ import "math"
 type DiceRoller struct {
 	diceCount  int
 	diceSides  int
-	totalRolls int
+	totalRolls uint64
 }
 
 type DiceRoll []int
@@ -14,7 +14,7 @@ func NewRoller(params DiceRollParameters) *DiceRoller {
 	return &DiceRoller{
 		diceCount:  params.DiceCount,
 		diceSides:  params.DiceSides,
-		totalRolls: int(math.Pow(float64(params.DiceSides), float64(params.DiceCount))),
+		totalRolls: uint64(math.Pow(float64(params.DiceSides), float64(params.DiceCount))),
 	}
 }
 
@@ -26,25 +26,25 @@ func (roller *DiceRoller) DiceSides() int {
 	return roller.diceSides
 }
 
-func (roller *DiceRoller) TotalRolls() int {
+func (roller *DiceRoller) TotalRolls() uint64 {
 	return roller.totalRolls
 }
 
-func (roller *DiceRoller) GetRollIdx(roll DiceRoll) int {
-	idx := 0
+func (roller *DiceRoller) GetRollIdx(roll DiceRoll) uint64 {
+	idx := uint64(0)
 	for _, dice := range roll {
-		idx = (roller.diceSides * idx) + int(dice) - 1
+		idx = (uint64(roller.diceSides) * idx) + uint64(dice) - 1
 	}
 	return idx
 }
 
-func (roller *DiceRoller) IdxToRoll(idx int) DiceRoll {
+func (roller *DiceRoller) IdxToRoll(idx uint64) DiceRoll {
 	dices := make(DiceRoll, roller.diceCount)
 	divisor := roller.totalRolls
 	residue := idx
 	for i := 0; i < roller.diceCount; i++ {
-		divisor /= roller.diceSides
-		dices[i] = (residue / divisor) + 1
+		divisor /= uint64(roller.diceSides)
+		dices[i] = int((residue / divisor) + 1)
 		residue %= divisor
 	}
 	return dices
